+++ exts/ansys.tools.omniverse.core/ansys/tools/omniverse/core/extension.py
@@ -1,27 +1,3 @@
-# Copyright (C) 2024 ANSYS, Inc. and/or its affiliates.
-# SPDX-License-Identifier: MIT
-#
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""Ansys Tools Omniverse Core Extension Module."""  # numpydoc ignore=EX01,SA01,ES01
-
 import json
 import logging
 import os
@@ -44,10 +20,8 @@ or by looking at CEI_HOME. CEI_HOME is tried first.
 """
 
 
-def find_kit_filename() -> Optional[str]:  # numpydoc ignore=SA01,EX01
+def find_kit_filename() -> Optional[str]:
     """
-    Find the location of the Kit executable.
-
     Use a combination of the current omniverse application and the information
     in the local .nvidia-omniverse/config/omniverse.toml file to come up with
     the pathname of a kit executable suitable for hosting another copy of the
@@ -55,8 +29,8 @@ def find_kit_filename() -> Optional[str]:  # numpydoc ignore=SA01,EX01
 
     Returns
     -------
-    str
-        The pathname of a kit executable or None.
+        The pathname of a kit executable or None
+
     """
     # get the current application
     app = omni.kit.app.get_app()
@@ -92,10 +66,8 @@ def find_kit_filename() -> Optional[str]:  # numpydoc ignore=SA01,EX01
     return None
 
 
-class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=PR01,SA01,EX01
+class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):
     """
-    AnsysToolsOmniverseCoreServerExtension extension class.
-
     This class is an Omniverse kit.  The kit is capable of creating a
     connection to an Ansys Distributed Scene Graph service and pushing
     the graph into an Omniverse Nucleus.
@@ -103,7 +75,7 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
 
     _service_instance = None
 
-    def __init__(self, *args, **kwargs) -> None:  # numpydoc ignore=GL08
+    def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
         ext_name = __name__.rsplit(".", 1)[0]
         self._logger = logging.getLogger(ext_name)
@@ -128,107 +100,103 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
         self._interpreter = self._find_ensight_cpython()
 
     @property
-    def version(self) -> str:  # numpydoc ignore=GL08
+    def version(self) -> str:
         return self._version
 
     @property
-    def dsg_uri(self) -> str:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """The endpoint of a Dynamic Scene Graph service:  grpc://{hostname}:{port}."""
+    def dsg_uri(self) -> str:
+        """The endpoint of a Dynamic Scene Graph service:  grpc://{hostname}:{port}"""
         return self._dsg_uri
 
     @dsg_uri.setter
-    def dsg_uri(self, uri: str) -> None:  # numpydoc ignore=GL08
+    def dsg_uri(self, uri: str) -> None:
         self._dsg_uri = uri
 
     @property
-    def destination(self) -> str:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """The output USD directory name."""
+    def destination(self) -> str:
+        """The output USD directory name"""
         return self._omni_uri
 
     @destination.setter
-    def destination(self, value: str) -> None:  # numpydoc ignore=GL08
+    def destination(self, value: str) -> None:
         self._omni_uri = value
 
     @property
-    def security_token(self) -> str:  # numpydoc ignore=ES01,RT01,SA01,EX01
+    def security_token(self) -> str:
         """The security token of the DSG service instance."""
         return self._security_token
 
     @security_token.setter
-    def security_token(self, value: str) -> None:  # numpydoc ignore=GL08
+    def security_token(self, value: str) -> None:
         self._security_token = value
 
     @property
-    def temporal(self) -> bool:  # numpydoc ignore=ES01,RT01,SA01,EX01
+    def temporal(self) -> bool:
         """If True, the DSG update should include all timesteps."""
         return self._temporal
 
     @temporal.setter
-    def temporal(self, value: bool) -> None:  # numpydoc ignore=GL08
+    def temporal(self, value: bool) -> None:
         self._temporal = bool(value)
 
     @property
-    def vrmode(self) -> bool:  # numpydoc ignore=ES01,RT01,SA01,EX01
+    def vrmode(self) -> bool:
         """If True, the DSG update should not include camera transforms."""
         return self._vrmode
 
     @vrmode.setter
-    def vrmode(self, value: bool) -> None:  # numpydoc ignore=GL08
+    def vrmode(self, value: bool) -> None:
         self._vrmode = bool(value)
 
     @property
-    def normalize_geometry(self) -> bool:  # numpydoc ignore=ES01,RT01,SA01,EX01
+    def normalize_geometry(self) -> bool:
         """If True, the DSG geometry should be remapped into normalized space."""
         return self._normalize_geometry
 
     @normalize_geometry.setter
-    def normalize_geometry(self, val: bool) -> None:  # numpydoc ignore=GL08
+    def normalize_geometry(self, val: bool) -> None:
         self._normalize_geometry = val
 
     @property
-    def time_scale(self) -> float:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """Value to multiply DSG time values by before passing to Omniverse."""
+    def time_scale(self) -> float:
+        """Value to multiply DSG time values by before passing to Omniverse"""
         return self._time_scale
 
     @time_scale.setter
-    def time_scale(self, value: float) -> None:  # numpydoc ignore=GL08
+    def time_scale(self, value: float) -> None:
         self._time_scale = value
 
     @property
-    def interpreter(self) -> str:  # numpydoc ignore=ES01,RT01,SA01,EX01
+    def interpreter(self) -> str:
         """Fully qualified path to the python.exe or .bat wrapper in which pyensight is installed."""
         return self._interpreter
 
     @interpreter.setter
-    def interpreter(self, value: str) -> None:  # numpydoc ignore=GL08
+    def interpreter(self, value: str) -> None:
         self._interpreter = value
 
     @classmethod
-    def get_instance(
-        cls,
-    ) -> Optional["AnsysToolsOmniverseCoreServerExtension"]:  # numpydoc ignore=GL08
+    def get_instance(cls) -> Optional["AnsysToolsOmniverseCoreServerExtension"]:
         return cls._service_instance
 
     @classmethod
-    def _setting(cls, name: str, env_varname: str = "") -> str:  # numpydoc ignore=SA01,EX01
+    def _setting(cls, name: str, env_varname: str = "") -> str:
         """
-        Get a CLI option value.
-
-        First check if any specified environment variable is present and if so, return that value.
+        Get a CLI option value. First check if any specified
+        environment variable is present and if so, return that value.
         Next, check to see if a command line value is set and return
         that.  Finally, fall back to the value (if any) specified in
         the kit toml file.
 
         Parameters
         ----------
-        name : str
+        name
             The name of the command line flag to check the value of.
-        env_varname : str
+        env_varname
             Optional name of the environment variable to check the value of.
 
         Returns
         -------
-        str
             A string or None.
         """
         # any environmental variable trumps them all.
@@ -241,49 +209,48 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
         s = f"/exts/{ext_name}/{name}"
         return settings.get(s)
 
-    def info(self, text: str) -> None:  # numpydoc ignore=ES01,SA01,EX01
+    def info(self, text: str) -> None:
         """
         Send message to the logger at the info level.
 
         Parameters
         ----------
-        text : str
+        text
             The message to send.
         """
         self._logger.info(text)
 
-    def warning(self, text: str) -> None:  # numpydoc ignore=ES01,SA01,EX01
+    def warning(self, text: str) -> None:
         """
         Send message to the logger at the warning level.
 
         Parameters
         ----------
-        text : str
+        text
             The message to send.
         """
         self._logger.warning(text)
 
-    def error(self, text: str) -> None:  # numpydoc ignore=ES01,SA01,EX01
+    def error(self, text: str) -> None:
         """
         Send message to the logger at the error level.
 
         Parameters
         ----------
-        text : str
+        text
             The message to send.
         """
         self._logger.error(text)
 
-    def _find_ensight_cpython(self) -> Optional[str]:  # numpydoc ignore=SA01,EX01
+    def _find_ensight_cpython(self) -> Optional[str]:
         """
         Scan the current system, looking for EnSight installations, specifically, cpython.
-
         Check: PYENSIGHT_ANSYS_INSTALLATION, CEI_HOME, AWP_ROOT* in that order
 
         Returns
         -------
-        str
-            The first cpython found or None.
+            The first cpython found or None
+
         """
 
         cpython = "cpython"
@@ -316,20 +283,20 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
                     return launch_file
         return ""
 
-    def on_startup(self, ext_id: str) -> None:  # numpydoc ignore=ES01,SA01,EX01
+    def on_startup(self, ext_id: str) -> None:
         """
         Called by Omniverse when the kit instance is started.
 
         Parameters
         ----------
-        ext_id : str
+        ext_id
             The specific version of the kit.
         """
         self._version = ext_id.split("-")[-1]
         self.info(f"ANSYS tools omniverse core server startup: {self._version}")
         AnsysToolsOmniverseCoreServerExtension._service_instance = self
 
-    def on_shutdown(self) -> None:  # numpydoc ignore=ES01,SA01,EX01
+    def on_shutdown(self) -> None:
         """
         Called by Omniverse when the kit instance is shutting down.
         """
@@ -337,20 +304,7 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
         self.shutdown()
         AnsysToolsOmniverseCoreServerExtension._service_instance = None
 
-    def validate_interpreter(self, launch_file: str) -> bool:  # numpydoc ignore=ES01,SA01,EX01
-        """
-        Validate the input interpreter.
-
-        Parameters
-        ----------
-        launch_file : str
-            The interpreter location.
-
-        Returns
-        -------
-        bool
-            True if the interpreter correctly has installed the omniverse cli tool.
-        """
+    def validate_interpreter(self, launch_file: str) -> bool:
         if len(launch_file) == 0:
             return False
         has_ov_module = False
@@ -365,10 +319,8 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
             has_ov_module = False
         return has_ov_module
 
-    def dsg_export(self) -> None:  # numpydoc ignore=SA01,EX01
+    def dsg_export(self) -> None:
         """
-        Export the current scene in USD format.
-
         Use the oneshot feature of the pyensight omniverse_cli to push the current
         EnSight scene to the supplied directory in USD format.
         """
@@ -404,7 +356,7 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
             self.warning(f"Error running translator: {error}")
         self._new_status_file(new=False)
 
-    def _new_status_file(self, new=True) -> None:  # numpydoc ignore=ES01,SA01,EX01
+    def _new_status_file(self, new=True) -> None:
         """
         Remove any existing status file and create a new one if requested.
 
@@ -425,17 +377,16 @@ class AnsysToolsOmniverseCoreServerExtension(omni.ext.IExt):  # numpydoc ignore=
                 tempfile.gettempdir(), str(uuid.uuid1()) + "_gs_status.txt"
             )
 
-    def read_status_file(self) -> dict:  # numpydoc ignore=ES01,SA01,EX01
-        """
-        Read the status file and return its contents as a dictionary.
+    def read_status_file(self) -> dict:
+        """Read the status file and return its contents as a dictionary.
 
         Note: this can fail if the file is being written to when this call is made, so expect
         failures.
 
         Returns
         -------
-        dict
-            A dictionary with the fields 'status', 'start_time', 'processed_buffers', 'total_buffers' or empty.
+        Optional[dict]
+            A dictionary with the fields 'status', 'start_time', 'processed_buffers', 'total_buffers' or empty
         """
         if not self._status_filename:
             return {}