+++ exts/ansys.tools.omniverse.dsgui/ansys/tools/omniverse/dsgui/extension.py
@@ -1,27 +1,3 @@
-# Copyright (C) 2024 ANSYS, Inc. and/or its affiliates.
-# SPDX-License-Identifier: MIT
-#
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""Ansys Tools Omniverse DSG UI Extension Module."""  # numpydoc ignore=EX01,SA01,ES01
-
 import logging
 import threading
 import time
@@ -32,16 +8,8 @@ import omni.ext
 import omni.ui as ui
 
 
-class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):  # numpydoc ignore=PR01,SA01,EX01
-    """
-    AnsysToolsOmniverseDSGUIExtension extension class.
-
-    This class is an Omniverse kit.  The kit is a GUI tool
-    that wraps the AnsysToolsOmniverseCoreServerExtension functionalities
-    and presents them to the user.
-    """
-
-    def __init__(self, *args, **kwargs) -> None:  # numpydoc ignore=GL08
+class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):
+    def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
         self._window: Any = None
         self._label_w: Any = None
@@ -61,22 +29,19 @@ class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):  # numpydoc ignore=PR01,
         self._error_msg = ""
 
     @property
-    def service(self) -> Optional["AnsysToolsOmniverseDSGUIExtension"]:  # numpydoc ignore=GL08
+    def service(self) -> Optional["AnsysToolsOmniverseDSGUIExtension"]:
         return ansys.tools.omniverse.core.AnsysToolsOmniverseCoreServerExtension.get_instance()
 
-    def info(self, text: str) -> None:  # numpydoc ignore=GL08
+    def info(self, text: str) -> None:
         self._logger.info(text)
 
-    def warning(self, text: str) -> None:  # numpydoc ignore=GL08
+    def warning(self, text: str) -> None:
         self._logger.warning(text)
 
-    def error(self, text: str) -> None:  # numpydoc ignore=GL08
+    def error(self, text: str) -> None:
         self._logger.error(text)
 
-    def start_server(self) -> None:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """
-        Start tge Ansys Tools Omniverse Core service.
-        """
+    def start_server(self) -> None:
         if self._connected:
             return
         self.service.dsg_uri = self._dsg_uri_w.model.as_string
@@ -93,16 +58,13 @@ class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):  # numpydoc ignore=PR01,
         self.info("Connected to DSG service")
         self._connected = True
 
-    def stop_server(self) -> None:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """
-        Stop the Ansys Tools Omniverse Core service.
-        """
+    def stop_server(self) -> None:
         if not self._connected:
             return
         self.info("Disconnect from DSG service")
         self._connected = False
 
-    def connect_cb(self) -> None:  # numpydoc ignore=GL08
+    def connect_cb(self) -> None:
         if self.service is None:
             self.error("Unable to find ansys.tools.omniverse.core instance")
             return
@@ -117,27 +79,24 @@ class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):  # numpydoc ignore=PR01,
                 self.start_server()
         self.update_ui()
 
-    def update_cb(self) -> None:  # numpydoc ignore=GL08
+    def update_cb(self) -> None:
         if not self._connected:
             self.error("No DSG service connected")
             return
         self.service.dsg_export()
 
-    def on_startup(self, ext_id: str) -> None:  # numpydoc ignore=GL08
+    def on_startup(self, ext_id: str) -> None:
         self.info(f"ANSYS tools omniverse DSG GUI startup: {ext_id}")
         if self.service is None:
             self.error("Unable to find ansys.tools.omniverse.core instance")
         self.build_ui()
         self._update_callback()
 
-    def _update_callback(self) -> None:  # numpydoc ignore=GL08
+    def _update_callback(self) -> None:
         self.update_ui()
         threading.Timer(0.5, self._update_callback).start()
 
-    def update_ui(self) -> None:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """
-        Update the UI depending on the export status.
-        """
+    def update_ui(self) -> None:
         status = self.service.read_status_file()
         if self._connected:
             self._connect_w.text = "Disconnect from DSG Server"
@@ -165,10 +124,7 @@ class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):  # numpydoc ignore=PR01,
         self._interpreter_w.enabled = not self._connected
         self._destination_w.enabled = not self._connected
 
-    def build_ui(self) -> None:  # numpydoc ignore=ES01,RT01,SA01,EX01
-        """
-        Build the extension UI.
-        """
+    def build_ui(self) -> None:
         self._window = ui.Window(f"ANSYS Tools Omniverse DSG ({self.service.version})")
         with self._window.frame:
             with ui.VStack(height=0, spacing=5):
@@ -239,7 +195,7 @@ class AnsysToolsOmniverseDSGUIExtension(omni.ext.IExt):  # numpydoc ignore=PR01,
                     self._connect_w = ui.Button("Connect to DSG Server", clicked_fn=self.connect_cb)
                     self._update_w = ui.Button("Request Update", clicked_fn=self.update_cb)
 
-    def on_shutdown(self) -> None:  # numpydoc ignore=GL08
+    def on_shutdown(self) -> None:
         self.info("ANSYS Tools Omniverse DSG shutdown")
         self.stop_server()
         self._window = None