+++ exts/ansys.tools.omniverse.core/docs/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0] - 2024-10-17
-- First release via ansys-tools-omnvierse repository
+## [1.0.0] - 2024-10-31
+- First release on ansys-tools-omniverse repo
 
 ## [0.8.12] - 2024-10-17
 - Update to let the path to the python installation be validated and edited