+++ exts/ansys.tools.omniverse.dsgui/docs/CHANGELOG.md
@@ -2,10 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0] - 2024-10-17
-- First release via ansys-tools-omnvierse repository
+## [1.0.0] - 2024-10-31
+- First release on ansys-tools-omniverse repo
 
 ## [0.8.12] - 2024-10-17
+- Update to let the path to the python installation be validated and edited
+
+## [0.8.11] - 2024-10-17
 - Add UI to view, edit, and validate the path to the python containing pyensight
 
 ## [0.8.11] - 2024-09-17